refactor: move settings to a dedicated file
